Prepare refactor of tasks into larger module
